requirements.txt
pymysql>=1.0.2
pikepdf>=6.0.0
ocrmypdf>=13.4.0
celery[redis]>=5.2.0
redis>=4.5.0
opensearch-py>=2.0.0
pdfminer.six>=20221105
python-dotenv>=1.0.0
